move ClientRequest,PlayerRequest to client::request module
