+++ spotify_player/src/event/mod.rs
@@ -1,4 +1,5 @@
 use crate::{
+    client::{ClientRequest, PlayerRequest},
     command::{self, Command},
     config,
     key::{Key, KeySequence},
@@ -14,73 +15,6 @@ use anyhow::{Context as _, Result};
 mod page;
 mod popup;
 mod window;
-
-#[derive(Debug)]
-/// A request that modifies the player's playback
-pub enum PlayerRequest {
-    NextTrack,
-    PreviousTrack,
-    Resume,
-    Pause,
-    ResumePause,
-    SeekTrack(chrono::Duration),
-    Repeat,
-    Shuffle,
-    Volume(u8),
-    ToggleMute,
-    TransferPlayback(String, bool),
-    StartPlayback(Playback, Option<bool>),
-}
-
-#[derive(Debug)]
-/// A request to the client
-pub enum ClientRequest {
-    GetCurrentUser,
-    GetDevices,
-    GetBrowseCategories,
-    GetBrowseCategoryPlaylists(Category),
-    GetUserPlaylists,
-    GetUserSavedAlbums,
-    GetUserFollowedArtists,
-    GetUserSavedTracks,
-    GetUserTopTracks,
-    GetUserRecentlyPlayedTracks,
-    GetContext(ContextId),
-    GetCurrentPlayback,
-    GetRadioTracks {
-        seed_uri: String,
-        seed_name: String,
-    },
-    Search(String),
-    AddTrackToQueue(TrackId<'static>),
-    AddTrackToPlaylist(PlaylistId<'static>, TrackId<'static>),
-    DeleteTrackFromPlaylist(PlaylistId<'static>, TrackId<'static>),
-    ReorderPlaylistItems {
-        playlist_id: PlaylistId<'static>,
-        insert_index: usize,
-        range_start: usize,
-        range_length: Option<usize>,
-        snapshot_id: Option<String>,
-    },
-    AddToLibrary(Item),
-    DeleteFromLibrary(ItemId),
-    ConnectDevice(Option<String>),
-    Player(PlayerRequest),
-    GetCurrentUserQueue,
-    #[cfg(feature = "lyric-finder")]
-    GetLyric {
-        track: String,
-        artists: String,
-    },
-    #[cfg(feature = "streaming")]
-    RestartIntegratedClient,
-    CreatePlaylist {
-        playlist_name: String,
-        public: bool,
-        collab: bool,
-        desc: String,
-    },
-}
 
 /// starts a terminal event handler (key pressed, mouse clicked, etc)
 pub fn start_event_handler(state: SharedState, client_pub: flume::Sender<ClientRequest>) {
